requirements.txt
anthropic
lark
pydantic
rich
python-dotenv